% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miPlot.R
\name{miPlot}
\alias{miPlot}
\title{Mi Plot}
\usage{
miPlot(MiData, type, y, col = "blue", bins = 30, by = "week",
  fill = "#575757")
}
\arguments{
\item{MiData}{is a list returned by function loadMiData}

\item{type}{character indicating type of plot. Options are "hist", "ts", "week", "box". "week" plots mean value of the same day in a week, such as mean value of Monday.}

\item{y}{variable to be plotted as y value. Options are "light", "deep", "sleep", "step". "Sleep" means to plot both light sleep and deep sleep as y values.}

\item{col}{color to be parsed to ggplot function.}

\item{bins}{only used in hist}
}
\value{
ggplot object
}
\description{
Generic plot function for Mi Data. Using rChart library to plot.
}
\author{
Borg
}

