% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepEfficiencyPlot.R
\name{sleepEfficiencyPlot}
\alias{sleepEfficiencyPlot}
\title{Sleep Efficiency Plot}
\usage{
sleepEfficiencyPlot(MiData, show = TRUE)
}
\arguments{
\item{MiData}{is a list returned by function loadMiData}

\item{show}{whether to show plots directly or just return plots}
}
\value{
a list of plots
}
\description{
Make plots of sleep efficiency against total sleep duration and date. Sleep efficiency is caculated as deep sleep duration divided by total sleep duration.
}
\author{
Borg
}

