% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepPlot.R
\name{sleepPlot}
\alias{sleepPlot}
\title{Sleep Plots}
\usage{
sleepPlot(MiData, show = TRUE)
}
\arguments{
\item{MiData}{is a list returned by function loadMiData}

\item{show}{whether to show plots directly or just return plots}
}
\value{
a list of plot objects
}
\description{
plots of light sleep and deep sleep showing trend and difference between light sleep and deep sleep.
}
\author{
Borg
}

