% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepPlot.R
\name{stepPlot}
\alias{stepPlot}
\title{Step Plots}
\usage{
stepPlot(MiData, show = TRUE)
}
\arguments{
\item{MiData}{is a list returned by function loadMiData}

\item{show}{whether to show plots directly or just return plots}
}
\value{
plot object
}
\description{
plot of step trend
}
\author{
Borg
}

