% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeksummary.R
\name{weeksummary}
\alias{weeksummary}
\title{weeksummary}
\usage{
weeksummary(MiData, show = TRUE)
}
\arguments{
\item{MiData}{is a list returned by function loadMiData}

\item{show}{whether to show plots directly or just return plots}
}
\value{
a list of plot objects
}
\description{
summary groupped by week
}
\author{
Borg
}

